+++ changelog.rst
@@ -5,6 +5,7 @@ Features:
 ---------
 
 * Add `max_field_width` setting to config, to enable more control over field truncation ([related issue](https://github.com/dbcli/pgcli/issues/1250)).
+* Re-run last query via bare `\watch`. (Thanks: `Saif Hakim`_)
 
 Bug fixes:
 ----------